pyproject.toml
[tool.poetry]
name = "wedge-golf-api"
version = "0.1.0"
description = "Wedge Golf API"
authors = ["Donato Di Ferdinando <fddiferd@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.95.0"
uvicorn = "^0.21.1"
gunicorn = "^20.1.0"
firebase-admin = "^6.1.0"
google-cloud-firestore = "^2.11.0"

[tool.poetry.dev-dependencies]
pytest = "^7.2.0"
black = "^23.1.0"
isort = "^5.11.4"
mypy = "^1.0.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api" 